Refactoring
removed escapingInput function and replaced it with the DOMPurify library
