+++ gaas-ui/package.json
@@ -34,6 +34,7 @@
     "core-js": "^3.21.1",
     "cors": "^2.8.5",
     "d3-array": "^3.1.1",
+    "dompurify": "^2.3.6",
     "env-cmd": "^10.1.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-prettier": "^4.0.0",
@@ -69,6 +70,7 @@
     "@testing-library/react": "^12.1.3",
     "@types/body-parser": "^1.19.2",
     "@types/core-js": "^2.5.5",
+    "@types/dompurify": "^2.3.3",
     "@types/enzyme": "^3.10.11",
     "@types/express": "^4.17.13",
     "@types/jest": "^27.4.1",
@@ -103,8 +105,8 @@
     "jsdom": "19.0.0",
     "jsdom-global": "3.0.2",
     "morgan": "^1.10.0",
-    "sass": "^1.49.9",
     "prettier": "^2.5.1",
+    "sass": "^1.49.9",
     "stylelint": "^14.5.3",
     "stylelint-config-standard": "^25.0.0",
     "stylelint-config-standard-scss": "^3.0.0",
