+++ gaas-ui/src/components/login/login-form.tsx
@@ -7,7 +7,7 @@ import { FormType } from "./login-modal";
 import { IAuthClient } from "../../rest/clients/authclient";
 import { AuthClientFactory } from "../../rest/clients/auth-client-factory";
 import { Copyright } from "../copyright/copyright";
-import escapingInputs from "../../util/sanitize-inputs";
+import DOMPurify from "dompurify";
 
 interface IProps {
     onChangeForm(fromType: FormType): void;
@@ -50,7 +50,7 @@ export default class LoginForm extends React.Component<IProps, IState> {
                 outcomeMessage: `Login failed: ${errorMessage}`,
             });
         };
-        this.authClient.login(escapingInputs(username), escapingInputs(password), onSuccess, onError);
+        this.authClient.login(DOMPurify.sanitize(username), DOMPurify.sanitize(password), onSuccess, onError);
     }
 
     public render() {
