+++ gaas-ui/src/util/util.js
@@ -1,0 +1,8 @@
+export function isJSONString(str) {
+    try {
+        JSON.parse(str);
+    } catch (e) {
+        return false;
+    }
+    return true;
+}
